bugfix and add example code
