--- ReadMe.md
@@ -1,10 +1,0 @@
-## これは何
-
-ContrasTech社のカメラをOpenFrameWorks上で動かせるようにしたやつ
-スレッド処理で450 ~ 500fpsくらい出せるようにした
-- http://www.contrastech.com/content/?875.html
-
-### 動作環境
-Linux/Windows
-
-- ofxCv & ofxOpenCv
