+++ example-viewer/addons.make
@@ -1,0 +1,2 @@
+ofxMarsCamera
+ofxOpenCv
