+++ libs/GenICam/include/define.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#define TRIGMODE_ON_STR             "On"            
+#define TRIGMODE_OFF_STR            "Off"           
+
+#define EXPOSUREAUTO_OFF_STR        "Off"           
+#define GAINAUTO_OFF_STR            "Off"           
+
+#define TRIGACT_FALLINGEDGE_STR     "FallingEdge"   
+#define TRIGACT_RISINGEDGE_STR      "RisingEdge"    
+
+#define TRIGSRC_SOFT_STR            "Software"		
+#define TRIGSRC_LINE1_STR           "Line1"
+
+#define IMAGE_SIZE					640 * 480
